requirements.txt
# Core Dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.3.0
joblib>=1.3.0

# ML Models
xgboost>=2.0.0
lightgbm>=4.0.0

# Gemini API
google-generativeai>=0.3.0

# UI/UX
rich>=13.0.0

# Configuration
pydantic>=2.0.0
